backend/docgen_service/app/utils/template_processor.py
"""
Template Processing Utilities
"""

from typing import Dict, Any
from docx import Document


class TemplateProcessor:
    """テンプレート処理ユーティリティ"""

    def process_template(self, template_path: str, data: Dict[str, Any]) -> Document:
        """テンプレートを処理"""
        doc = Document(template_path)
        # テンプレート処理ロジック
        return doc
